feat: Change button background color from black/20 to white/10 in AttendeeList component
This diff changes the background color of several buttons in the AttendeeList component from a darker shade (black/20) to a lighter shade (white/10). The buttons affected include the "More" button in each attendee row, as well as the pagination buttons at the bottom of the list. The border color of these buttons is also changed to match the new background color.