feat: "Add props to TableCell and update element from 'td' to 'th'"
Explanation:
The code diff shows that the `TableCell` component's props have been updated to extend `ComponentProps<"td">` which allows for additional props to be passed to the component. The component's implementation has also been updated to use the 'th' element instead of 'td', and the 'children' are now being passed the additional props. A class name has also been added to the 'th' element.