feat: "Disable pagination buttons when on the first/last page in AttendeeList"
This commit disables the "First Page" and "Previous Page" buttons when on the first page, and disables the "Next Page" and "Last Page" buttons when on the last page in the AttendeeList component. This prevents users from clicking on these buttons when it is not possible to change the page.